Moved functions into Grid.js
