+++ client/src/components/dashboard.js
@@ -1,13 +1,13 @@
 import React from "react";
 import RevolverImage from "../images/revolver.png"
 
-function DropdownMenu() {
+function Dashboard() {
 
   return (
     <div className="dashboard">
-      <img className="revolver" src={RevolverImage}></img>
+      <img className="revolver" src={RevolverImage} alt="revolverImg"></img>
     </div>
   )
 }
 
-export default DropdownMenu;
+export default Dashboard;
